P9_03_flask_app/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform



pillow
tensorflow==2.7.0
numpy
pandas
flask
gunicorn
pillow
opencv-python
